requirements.txt
edge-tts>=6.1.0
google-genai>=1.0.0
pyyaml>=6.0